+++ R/LearnerTorchModule.R
@@ -27,6 +27,7 @@
 #' @param properties (`NULL` or `character()`)\cr
 #'   The properties of the learner.
 #'   Defaults to all available properties for the given task type.
+#' @template param_predict_types
 # @section Parameters: See [`LearnerTorch`] and constructor argument `param_set`.
 #' @family Learner
 #' @include LearnerTorch.R
@@ -35,7 +36,7 @@
 #' nn_one_layer = nn_module("nn_one_layer",
 #'   initialize = function(task, size_hidden) {
 #'     self$first = nn_linear(task$n_features, size_hidden)
-#'     self$second = nn_linear(size_hidden, length(task$class_names))
+#'     self$second = nn_linear(size_hidden, output_dim_for(task))
 #'   },
 #'   # argument x corresponds to the ingress token x
 #'   forward = function(x) {
@@ -61,7 +62,10 @@ LearnerTorchModule = R6Class("LearnerTorchModule",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function(module_generator = NULL, param_set = NULL, ingress_tokens = NULL,
       task_type, properties = NULL, optimizer = NULL, loss = NULL, callbacks = list(),
-      packages = character(0), feature_types = NULL) {
+      packages = character(0), feature_types = NULL, predict_types = NULL) {
+      if (is.null(task_type)) {
+        stopf("task_type must be provided")
+      }
       assert(check_class(module_generator, "nn_module_generator"), check_function(module_generator))
       private$.module_generator = module_generator
       args = names(formals(module_generator))
@@ -69,7 +73,7 @@ LearnerTorchModule = R6Class("LearnerTorchModule",
         stopf("module_generator must have 'task' as a parameter")
       }
 
-      private$.ingress_tokens = assert_list(ingress_tokens, types = "TorchIngressToken", names = "unique", min.len = 1L)
+      private$.ingress_tokens_ = assert_list(ingress_tokens, types = "TorchIngressToken", names = "unique", min.len = 1L)
 
       if (is.null(feature_types)) {
         feature_types = unname(mlr_reflections$task_feature_types)
@@ -99,12 +103,13 @@ LearnerTorchModule = R6Class("LearnerTorchModule",
         packages = packages,
         param_set = param_set,
         feature_types = feature_types,
+        predict_types = predict_types,
         man = "mlr3torch::mlr_learners.module"
       )
     }
   ),
   private = list(
-    .ingress_tokens = NULL,
+    .ingress_tokens_ = NULL,
     .module_generator = NULL,
 
     .network = function(task, param_vals) {
@@ -112,13 +117,8 @@ LearnerTorchModule = R6Class("LearnerTorchModule",
       invoke(private$.module_generator, task = task, .args = module_params)
     },
 
-    .dataset = function(task, param_vals) {
-      ingress_tokens = private$.ingress_tokens
-      dataset = task_dataset(
-        task,
-        feature_ingress_tokens = ingress_tokens,
-        target_batchgetter = get_target_batchgetter(self$task_type)
-      )
+    .ingress_tokens = function(task, param_vals) {
+      private$.ingress_tokens_
     },
 
     .additional_phash_input = function() {