+++ tests/testthat/test_LearnerTorchTabResNet.R
@@ -46,3 +46,27 @@ test_that("cloning also keeps parameter values", {
   expect_equal(learner$param_set$values$n_blocks, 2)
   expect_equal(learnerc$param_set$values$n_blocks, 2)
 })
+
+test_that("task types", {
+  learner = lrn("classif.tab_resnet", n_blocks = 0, epochs = 0, batch_size = 16, d_block = 5, d_hidden = 10, dropout1 = 0.3, dropout2 = 0.3)
+  expect_learner_torch(learner, tsk("iris"))
+  expect_learner_torch(learner, tsk("sonar"))
+  learner = lrn("regr.tab_resnet", n_blocks = 0, epochs = 0, batch_size = 16, d_block = 5, d_hidden = 10, dropout1 = 0.3, dropout2 = 0.3)
+  expect_learner_torch(learner, tsk("mtcars"))
+})
+
+
+test_that("lazy tensor works", {
+  learner = lrn("classif.tab_resnet", n_blocks = 1, epochs = 1, batch_size = 16, d_block = 5, d_hidden = 10, dropout1 = 0.3, dropout2 = 0.3)
+  task = tsk("lazy_iris")$filter(c(1, 51))
+  expect_error(learner$train(task), regexp = NA)
+})
+
+
+test_that("error messages", {
+  l = lrn("classif.tab_resnet", n_blocks = 1, epochs = 0, batch_size = 16, d_block = 5, d_hidden = 10, dropout1 = 0.3, dropout2 = 0.3)
+  expect_error(l$train(nano_imagenet()), "expects an input shape of length 2")
+  expect_error(l$train(nano_dogs_vs_cats()), "Please specify the learner's `shape` parameter")
+  l$configure(shape = c(NA, 4))
+  expect_error(l$train(nano_dogs_vs_cats()), NA)
+})